Downloader now uses floats for progress properly. Fixes issue 1404
git-svn-id: http://svn.dmdirc.com/trunk@4300 00569f92-eb28-0410-84fd-f71c24880f43

+++ src/com/dmdirc/util/Downloader.java
@@ -171,7 +171,7 @@ public final class Downloader {
                 output.write(buffer, 0, count);
                 
                 if (listener != null) {
-                    listener.downloadProgress(100 * current / length);
+                    listener.downloadProgress(100 * (float) current / length);
                 }
             }
         } while (count > 0);
